requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
ffmpy==0.3.0
idna==3.3
m3u-ipytv==0.1.3
pkg-resources==0.0.0
progressbar2==4.0.0
python-utils==3.3.3
requests==2.28.1
termcolor==1.1.0
urllib3==1.26.12
